feat: webdav-sync-hook for termux-file-editor
termux offers itself to accept shared files on android. Those files are
then used by a script under `$HOME/bin/termux-file-editor` as `$1`.
The webdav-sync-hook allows to rename this file and place it a location
on the registered `$WEBDAV_URL`-server.
Installation happens regularly, if no other `termux-file-editor` was
registered yet.